DPR-320 factored out json validator into separate class
